+++ jax/experimental/mosaic/gpu/__init__.py
@@ -71,7 +71,7 @@ from .fragmented_array import (
     WGMMA_LAYOUT_UPCAST_2X as WGMMA_LAYOUT_UPCAST_2X,
     WGMMA_LAYOUT_UPCAST_4X as WGMMA_LAYOUT_UPCAST_4X,
     TMEM_NATIVE_LAYOUT as TMEM_NATIVE_LAYOUT,
-    TMA_GATHER_INDICES_LAYOUT as TMA_GATHER_INDICES_LAYOUT,
+    TMA_INDICES_LAYOUT as TMA_INDICES_LAYOUT,
     tmem_native_layout as tmem_native_layout,
     WGSplatFragLayout as WGSplatFragLayout,
     WGStridedFragLayout as WGStridedFragLayout,